refactoring and max tip index
